refactor: change structure of the project, edit styles
