+++ src/assets/style/main.scss
@@ -300,7 +300,13 @@ footer {
   gap: 6px;
   border-radius: 10px;
   margin-bottom: 30px;
-  // background: black;
+  transition: all 0.6s ease-in-out;
+  &--disabled {
+    opacity: 0.4;
+    &:hover {
+      opacity: 1;
+    }
+  }
 }
 
 .status {
@@ -324,7 +330,7 @@ footer {
 
 .square:hover,
 .square:focus {
-  background: #ddd;
+  background: #e0e0e0;
 }
 
 .square:focus {
@@ -342,12 +348,73 @@ footer {
   align-items: center;
   text-align: center;
   width: 100%;
-  height: 80vh;
+  height: calc(100vh - 200px);
   flex-direction: row;
 }
 
 .game-info {
   margin-left: 20px;
+}
+
+.status {
+  font-size: 20px;
+  font-weight: bold;
+  color: #009688;
+  margin-bottom: 20px;
+  transition: all 0.4s ease;
+  &--win {
+    color: #ed4b82;
+  }
+}
+
+/* ------------------------------ ADDITIONAL STYLES ----------------------------- */
+/* ------------------------------ 0 ANIMATION ----------------------------- */
+.circle,
+.first,
+.second {
+  opacity: 0;
+  transition: all 0.3s ease-in-out;
+}
+
+.circle {
+  display: block;
+  width: 50px;
+  height: 50px;
+  stroke-dasharray: 283;
+  stroke-dashoffset: 283;
+  -webkit-animation: draw 1s ease-in-out forwards;
+  animation: draw 1s ease-in-out forwards;
+  margin: auto;
+}
+
+/* ------------------------------ X ANIMATION ----------------------------- */
+.cross {
+  display: block;
+  width: 45px;
+  height: 45px;
+  margin: auto;
+}
+
+.first {
+  stroke-dasharray: 125;
+  stroke-dashoffset: 125;
+  -webkit-animation: draw 0.6s forwards;
+  animation: draw 0.6s forwards;
+}
+
+.second {
+  stroke-dasharray: 125;
+  stroke-dashoffset: 125;
+  -webkit-animation: draw 0.6s 0.6s forwards;
+  animation: draw 0.6s 0.6s forwards;
+}
+
+/* ------------------------------ X ANIMATION ----------------------------- */
+@keyframes draw {
+  100% {
+    stroke-dashoffset: 0;
+    opacity: 1;
+  }
 }
 
 @import './media';
