+++ src/containers/pages/Board.tsx
@@ -1,23 +1,9 @@
 import {Button} from '@material-ui/core';
 import React, {FC, useState} from 'react';
-import {Fade} from '@material-ui/core';
-
-interface IProps {
-  value: string;
-  setActive: () => void;
-}
-
-// Square
-export const Square: FC<IProps> = ({value, setActive}) => {
-  return (
-    <button className="square" onClick={setActive}>
-      {value}
-    </button>
-  );
-};
+import {Square} from '../../components/Square';
 
 // Board - first child
-export const Board = () => {
+export const Board: FC = () => {
   const [squares, setSquares] = useState<Array<string>>(Array(9).fill(null));
   const [xIsNext, setXIsNext] = useState<boolean>(true);
   const [winner, setWinner] = useState<string | null>(null);
@@ -35,7 +21,7 @@ export const Board = () => {
       [2, 5, 8],
       // diagonal indexes
       [0, 4, 8],
-      [2, 4, 6],
+      [2, 4, 6]
     ];
 
     for (let i = 0; i < winnerLines.length; i++) {
@@ -78,7 +64,9 @@ export const Board = () => {
     // this.audio = new Audio('audio_url');
     // this.audio.play();
 
-    const audio = new Audio('https://s3.amazonaws.com/freecodecamp/drums/Heater-1.mp3');
+    const audioUrl = xIsNext ? 'https://timra.ru/portfolio/audio/cross.mp3' : 'https://timra.ru/portfolio/audio/zero.mp3';
+    const audio = new Audio(audioUrl);
+    audio.volume = 0.1;
     audio.play();
 
     setSquares(array);
@@ -89,40 +77,23 @@ export const Board = () => {
 
   const handlePlayAgain = () => {
     setSquares(Array(9).fill(null));
+    setWinner(null);
     setIsTie(false);
   };
 
   return (
-    <Fade in={true} timeout={200}>
-      <>
-        <div className="status">&nbsp;{getStatus(getWinner(squares), xIsNext)}</div>
-        <div className="board">
-          {squares.map((el, idx) => {
-            return <Square key={idx} value={el} setActive={() => handleClick(idx)} />;
-          })}
-        </div>
-        {(winner || isTie) && (
-          <Button variant="contained" onClick={handlePlayAgain}>
-            Play again
-          </Button>
-        )}
-      </>
-    </Fade>
-  );
-};
-
-// Game - parent
-export const Game = () => {
-  return (
-    <div className="game">
-      <div className="game-board">
-        <h1>Tic-Tac-Toe</h1>
-        <Board />
+    <>
+      <div className={winner || isTie ? 'status status--win' : 'status'}>&nbsp;{getStatus(getWinner(squares), xIsNext)}</div>
+      <div className={winner || isTie ? 'board board--disabled' : 'board'}>
+        {squares.map((el, idx) => {
+          return <Square key={idx} value={el} setActive={() => handleClick(idx)} />;
+        })}
       </div>
-      <div className="game-info">
-        <div>{/* status */}</div>
-        <ol>{/* TODO */}</ol>
-      </div>
-    </div>
+      {(winner || isTie) && (
+        <Button variant="contained" onClick={handlePlayAgain}>
+          Play again
+        </Button>
+      )}
+    </>
   );
 };
